refactor: #138 logout refactoring
