+++ src/main/java/com/aliens/friendship/domain/auth/repository/RefreshTokenRepository.java
@@ -9,6 +9,7 @@ public interface RefreshTokenRepository
         extends MongoRepository<RefreshToken, String> {
 
     boolean existsByEmail(String email);
-
     Optional<RefreshToken> findByEmailAndValue(String email, String value);
+
+    void deleteAllByEmail(String email);
 }
